-Wall

+++ src/goal_types.cpp
@@ -145,7 +145,8 @@ void TouchGoal::describe(GoalContext& context) const {
             new collision_detection::FCLGeometry(fcl, link_model, shape_index));
         s.edges.resize(s.points.size());
         std::vector<std::unordered_set<size_t>> edge_sets(s.points.size());
-        for (size_t edge_index = 0; edge_index < fcl->num_edges; edge_index++) {
+        for (size_t edge_index = 0;
+             edge_index < static_cast<size_t>(fcl->num_edges); edge_index++) {
           auto edge = fcl->edges[edge_index];
           if (edge_sets[edge.first].find(edge.second) ==
               edge_sets[edge.first].end()) {
